Fix MVC integration with UseEndpoint (#8047)
